requirements.txt
azure-core==1.20.0
azure-storage-blob==12.9.0
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.4

